Stacktraces
This code handles exceptions horribly - honastly, that in that aspect
this is the worst. You took everything that could possibly go wrong,
caught it, and returned false. How did you debug when something went
wrong? How did you know something went wrong?

You do know that RuntimeExceptions are never suppost to happen - if they
do, it's a programming error. Excpetions, on the other hand, need to be
caught, because the programmer can't program something that will never
throw that exception.

Ex. ArrayIndexOutOfBoundsException
This extends RuntimeException because Java assumes that you, as a Java
dev, know to not make calls with bad indexes.

An exception like IOException is thrown when an IO error happens, which
is compleatly out of control of the programmer. Java forces you to deal
with it.

catching runtime exceptions is horrible code - you should never catch
anything that extends RuntimeException.

Exception is even worse, and Throwable will catch even Errors

W H Y